gdk-wayland: Manually implement Toplevel.export_handle
To convert a NULL handle into an error
Fixes #1565
