+++ gdk4-wayland/Gir.toml
@@ -14,8 +14,6 @@ trust_return_value_nullability = true
 generate = [
     "GdkWayland.WaylandGLContext",
     "GdkWayland.WaylandPopup",
-    "GdkWayland.WaylandToplevel",
-    "GdkWayland.WaylandToplevelExported"
 ]
 
 manual = [
@@ -108,4 +106,11 @@ manual_traits = ["WaylandSurfaceExtManual"]
     [[object.function]]
     name = "get_wl_surface"
     rename = "wl_surface"
+    manual = true
+
+[[object]]
+name = "GdkWayland.WaylandToplevel"
+status = "generate"
+    [[object.function]]
+    name = "export_handle"
     manual = true
