+++ gdk4-wayland/src/auto/wayland_toplevel.rs
@@ -5,7 +5,6 @@
 
 use crate::WaylandSurface;
 use glib::translate::*;
-use std::boxed::Box as Box_;
 
 glib::wrapper! {
     #[doc(alias = "GdkWaylandToplevel")]
@@ -26,37 +25,6 @@ impl WaylandToplevel {
                 self.to_glib_none().0,
                 handle.to_glib_none().0,
             );
-        }
-    }
-
-    #[doc(alias = "gdk_wayland_toplevel_export_handle")]
-    pub fn export_handle<P: Fn(&WaylandToplevel, &str) + 'static>(&self, callback: P) -> bool {
-        let callback_data: Box_<P> = Box_::new(callback);
-        unsafe extern "C" fn callback_func<P: Fn(&WaylandToplevel, &str) + 'static>(
-            toplevel: *mut ffi::GdkWaylandToplevel,
-            handle: *const libc::c_char,
-            user_data: glib::ffi::gpointer,
-        ) {
-            let toplevel = from_glib_borrow(toplevel);
-            let handle: Borrowed<glib::GString> = from_glib_borrow(handle);
-            let callback = &*(user_data as *mut P);
-            (*callback)(&toplevel, handle.as_str())
-        }
-        let callback = Some(callback_func::<P> as _);
-        unsafe extern "C" fn destroy_func_func<P: Fn(&WaylandToplevel, &str) + 'static>(
-            data: glib::ffi::gpointer,
-        ) {
-            let _callback = Box_::from_raw(data as *mut P);
-        }
-        let destroy_call3 = Some(destroy_func_func::<P> as _);
-        let super_callback0: Box_<P> = callback_data;
-        unsafe {
-            from_glib(ffi::gdk_wayland_toplevel_export_handle(
-                self.to_glib_none().0,
-                callback,
-                Box_::into_raw(super_callback0) as *mut _,
-                destroy_call3,
-            ))
         }
     }
